shared dateTime utils, restructure SchedulePromptModal
