+++ ui/client/src/components/shared/NotificationPanel.tsx
@@ -5,6 +5,7 @@ import { Badge } from '@/components/ui/badge';
 import SimpleTooltip from '@/components/shared/SimpleTooltip';
 import { useNotifications } from '@/contexts/NotificationContext';
 import { ShareInvite, formatShareSenderLabel } from '@/api/shares';
+import { formatRelativeTime } from '@/utils/dateTimeUtils';
 
 interface NotificationPanelProps {
   isOpen: boolean;
@@ -59,18 +60,8 @@ export default function NotificationPanel({ isOpen, onClose }: NotificationPanel
   };
 
   // Format date for display
-  const formatDate = (dateString: string) => {
-    const date = new Date(dateString);
-    const now = new Date();
-    const diffInMinutes = Math.floor((now.getTime() - date.getTime()) / (1000 * 60));
-    
-    if (diffInMinutes < 1) return 'Just now';
-    if (diffInMinutes < 60) return `${diffInMinutes}m ago`;
-    if (diffInMinutes < 1440) return `${Math.floor(diffInMinutes / 60)}h ago`;
-    if (diffInMinutes < 10080) return `${Math.floor(diffInMinutes / 1440)}d ago`;
-    
-    return date.toLocaleDateString();
-  };
+  const formatDate = (dateString: string) =>
+    formatRelativeTime(dateString, { justNowLabel: 'Just now', capDays: 7 });
 
   // Filter only pending notifications for notifications panel
   const pendingNotifications = receivedNotifications?.filter(notification => notification.status === 'pending');
