+++ ui/client/src/utils/index.ts
@@ -1,3 +1,5 @@
+import { formatDateTime, formatRelativeTime } from "./dateTimeUtils";
+
 /**
  * Format numbers with K suffix for large numbers
  */
@@ -14,34 +16,14 @@ export function formatNumber(num: number | string): string {
  */
 export function getLastSyncTime(lastSyncAt?: string): string {
   if (!lastSyncAt) return "Never";
-  const lastSync = new Date(lastSyncAt);
-  const now = new Date();
-  const diffMinutes = Math.floor((now.getTime() - lastSync.getTime()) / 60000);
-
-  if (diffMinutes < 1) return "just now";
-  if (diffMinutes < 60) return `${diffMinutes}m ago`;
-
-  const diffHours = Math.floor(diffMinutes / 60);
-  if (diffHours < 24) return `${diffHours}h ago`;
-
-  const diffDays = Math.floor(diffHours / 24);
-  return `${diffDays}d ago`;
+  return formatRelativeTime(lastSyncAt);
 };
 
 /**
  * Get time ago string with more granular options
  */
 export function timeAgo(dateStr: string): string {
-  const now = new Date();
-  const past = new Date(dateStr);
-  const minutes = Math.floor((+now - +past) / 60000);
-
-  if (minutes < 1) return "just now";
-  if (minutes < 60) return `${minutes} minute${minutes > 1 ? "s" : ""} ago`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours} hour${hours > 1 ? "s" : ""} ago`;
-  const days = Math.floor(hours / 24);
-  return `${days} day${days > 1 ? "s" : ""} ago`;
+  return formatRelativeTime(dateStr, { verbose: true });
 };
 
 /**
@@ -49,15 +31,7 @@ export function timeAgo(dateStr: string): string {
  */
 export function formatDate(dateStr: string): string {
   if (!dateStr) return '';
-  const date = new Date(dateStr);
-  return date.toLocaleString('en-GB', {
-    day: '2-digit',
-    month: 'short',
-    year: 'numeric',
-    hour: '2-digit',
-    minute: '2-digit',
-    hour12: false
-  });
+  return formatDateTime(dateStr);
 };
 
 /**
@@ -65,16 +39,5 @@ export function formatDate(dateStr: string): string {
  * Falls back to date string for timestamps older than 7 days
  */
 export function formatRelativeTimestamp(timestamp: string): string {
-  const date = new Date(timestamp);
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffMins = Math.floor(diffMs / 60000);
-  const diffHours = Math.floor(diffMs / 3600000);
-  const diffDays = Math.floor(diffMs / 86400000);
-
-  if (diffMins < 1) return "Just now";
-  if (diffMins < 60) return `${diffMins}m ago`;
-  if (diffHours < 24) return `${diffHours}h ago`;
-  if (diffDays < 7) return `${diffDays}d ago`;
-  return date.toLocaleDateString();
+  return formatRelativeTime(timestamp, { justNowLabel: "Just now", capDays: 7 });
 };
